Adjusted DbBindableLong to be 64 bits if needed and adjusted timing for 64 bits

+++ engine/server/application/TaskManager/src/shared/ManagerConnection.cpp
@@ -122,8 +122,8 @@ void ManagerConnection::onReceive(const Archive::ByteStream & message)
 		}
 		case constcrc("TaskConnectionIdMessage") :
 		{
-			static long const clockDriftFatalTimePeriod = static_cast<long>(TaskManager::getStartTime()) + ConfigTaskManager::getClockDriftFatalIntervalSeconds();
-			long const currentTime = static_cast<long>(::time(nullptr));
+			static uint32_t const clockDriftFatalTimePeriod = static_cast<uint32_t>(TaskManager::getStartTime()) + ConfigTaskManager::getClockDriftFatalIntervalSeconds();
+			uint32_t const currentTime = static_cast<uint32_t>(::time(nullptr));
 			TaskConnectionIdMessage t(r);
 			WARNING_STRICT_FATAL(t.getServerType() != TaskConnectionIdMessage::TaskManager,
 								 ("ManagerConnection received wrong type identifier"));
